+++ packages/create-instance/create-functional-component.js
@@ -2,7 +2,8 @@
 
 import { throwError } from 'shared/util'
 import { validateSlots } from './validate-slots'
-import { createSlotVNodes } from './add-slots'
+import { createSlotVNodes } from './create-slot-vnodes'
+import createScopedSlots from './create-scoped-slots'
 
 export default function createFunctionalComponent (
   component: Component,
@@ -15,17 +16,21 @@ export default function createFunctionalComponent (
     validateSlots(mountingOptions.slots)
   }
 
+  const data = mountingOptions.context ||
+    component.FunctionalRenderContext || {}
+  data.scopedSlots = createScopedSlots(mountingOptions.scopedSlots)
+
   return {
     render (h: Function) {
       return h(
         component,
-        mountingOptions.context || component.FunctionalRenderContext,
+        data,
         (mountingOptions.context &&
           mountingOptions.context.children &&
           mountingOptions.context.children.map(
             x => (typeof x === 'function' ? x(h) : x)
           )) ||
-          createSlotVNodes(h, mountingOptions.slots || {})
+          createSlotVNodes(this, mountingOptions.slots || {})
       )
     },
     name: component.name,