+++ packages/create-instance/create-instance.js
@@ -1,17 +1,29 @@
 // @flow
 
-import { createSlotVNodes } from './add-slots'
+import { createSlotVNodes } from './create-slot-vnodes'
 import addMocks from './add-mocks'
 import { addEventLogger } from './log-events'
 import { createComponentStubs } from 'shared/stub-components'
 import { throwError, vueVersion } from 'shared/util'
 import { compileTemplate } from 'shared/compile-template'
+import { isRequiredComponent } from 'shared/validators'
 import extractInstanceOptions from './extract-instance-options'
 import createFunctionalComponent from './create-functional-component'
-import { componentNeedsCompiling } from 'shared/validators'
+import { componentNeedsCompiling, isPlainObject } from 'shared/validators'
 import { validateSlots } from './validate-slots'
 import createScopedSlots from './create-scoped-slots'
 import { extendExtendedComponents } from './extend-extended-components'
+
+function compileTemplateForSlots (slots: Object): void {
+  Object.keys(slots).forEach(key => {
+    const slot = Array.isArray(slots[key]) ? slots[key] : [slots[key]]
+    slot.forEach(slotValue => {
+      if (componentNeedsCompiling(slotValue)) {
+        compileTemplate(slotValue)
+      }
+    })
+  })
+}
 
 export default function createInstance (
   component: Component,
@@ -51,8 +63,20 @@ export default function createInstance (
 
   addEventLogger(_Vue)
 
+  // Replace globally registered components with components extended
+  // from localVue. This makes sure the beforeMount mixins to add stubs
+  // is applied to globally registered components.
+  // Vue version must be 2.3 or greater, because of a bug resolving
+  // extended constructor options (https://github.com/vuejs/vue/issues/4976)
+  if (vueVersion > 2.2) {
+    for (const c in _Vue.options.components) {
+      if (!isRequiredComponent(c)) {
+        _Vue.component(c, _Vue.extend(_Vue.options.components[c]))
+      }
+    }
+  }
+
   const stubComponents = createComponentStubs(
-    // $FlowIgnore
     component.components,
     // $FlowIgnore
     options.stubs
@@ -66,21 +90,46 @@ export default function createInstance (
   )
 
   // stub components should override every component defined in a component
-  _Vue.mixin({
-    created () {
-      Object.assign(
-        this.$options.components,
-        stubComponents
-      )
+  if (options.stubs) {
+    instanceOptions.components = {
+      ...instanceOptions.components,
+      ...stubComponents
     }
+  }
+  function addStubComponentsMixin () {
+    Object.assign(
+      this.$options.components,
+      stubComponents
+    )
+  }
+  _Vue.mixin({
+    beforeMount: addStubComponentsMixin,
+    // beforeCreate is for components created in node, which
+    // never mount
+    beforeCreate: addStubComponentsMixin
   })
 
   if (component.options) {
     component.options._base = _Vue
   }
 
-  const Constructor = vueVersion < 2.3 && typeof component === 'function'
-    ? component.extend(instanceOptions)
+  function getExtendedComponent (component, instanceOptions) {
+    const extendedComponent = component.extend(instanceOptions)
+    // to keep the possible overridden prototype and _Vue mixins,
+    // we need change the proto chains manually
+    // @see https://github.com/vuejs/vue-test-utils/pull/856
+    // code below equals to
+    // `extendedComponent.prototype.__proto__.__proto__ = _Vue.prototype`
+    const extendedComponentProto =
+      Object.getPrototypeOf(extendedComponent.prototype)
+    Object.setPrototypeOf(extendedComponentProto, _Vue.prototype)
+
+    return extendedComponent
+  }
+
+  // extend component from _Vue to add properties and mixins
+  const Constructor = typeof component === 'function'
+    ? getExtendedComponent(component, instanceOptions)
     : _Vue.extend(component).extend(instanceOptions)
 
   Constructor._vueTestUtilsRoot = component
@@ -90,6 +139,8 @@ export default function createInstance (
   })
 
   if (options.slots) {
+    compileTemplateForSlots(options.slots)
+    // $FlowIgnore
     validateSlots(options.slots)
   }
 
@@ -106,25 +157,36 @@ export default function createInstance (
 
   const scopedSlots = createScopedSlots(options.scopedSlots)
 
-  const Parent = _Vue.extend({
-    provide: options.provide,
-    render (h) {
-      const slots = options.slots
-        ? createSlotVNodes(h, options.slots)
-        : undefined
-      return h(
-        Constructor,
-        {
-          ref: 'vm',
-          props: options.propsData,
-          on: options.listeners,
-          attrs: options.attrs,
-          scopedSlots
+  if (options.parentComponent && !isPlainObject(options.parentComponent)) {
+    throwError(
+      `options.parentComponent should be a valid Vue component ` +
+      `options object`
+    )
+  }
+
+  const parentComponentOptions = options.parentComponent || {}
+  parentComponentOptions.provide = options.provide
+  parentComponentOptions.render = function (h) {
+    const slots = options.slots
+      ? createSlotVNodes(this, options.slots)
+      : undefined
+    return h(
+      Constructor,
+      {
+        ref: 'vm',
+        on: options.listeners,
+        attrs: {
+          ...options.attrs,
+          // pass as attrs so that inheritAttrs works correctly
+          // propsData should take precedence over attrs
+          ...options.propsData
         },
-        slots
-      )
-    }
-  })
+        scopedSlots
+      },
+      slots
+    )
+  }
+  const Parent = _Vue.extend(parentComponentOptions)
 
   return new Parent()
 }