+++ packages/test-utils/src/create-wrapper.js
@@ -6,7 +6,7 @@ import VueWrapper from './vue-wrapper'
 
 export default function createWrapper (
   node: VNode | Component,
-  options: WrapperOptions
+  options: WrapperOptions = {}
 ): VueWrapper | Wrapper {
   const componentInstance = node.componentInstance || node.child
   if (componentInstance) {