+++ packages/test-utils/src/index.js
@@ -4,6 +4,9 @@ import createLocalVue from './create-local-vue'
 import TransitionStub from './components/TransitionStub'
 import TransitionGroupStub from './components/TransitionGroupStub'
 import RouterLinkStub from './components/RouterLinkStub'
+import createWrapper from './create-wrapper'
+import Wrapper from './wrapper'
+import WrapperArray from './wrapper-array'
 import config from './config'
 import { warn } from 'shared/util'
 
@@ -17,11 +20,14 @@ function shallow (component, options) {
 
 export default {
   createLocalVue,
+  createWrapper,
   config,
   mount,
   shallow,
   shallowMount,
   TransitionStub,
   TransitionGroupStub,
-  RouterLinkStub
+  RouterLinkStub,
+  Wrapper,
+  WrapperArray
 }