+++ test/specs/mounting-options/attachToDocument.spec.js
@@ -1,11 +1,15 @@
-import { compileToFunctions } from 'vue-template-compiler'
 import { describeWithShallowAndMount, isRunningJSDOM } from '~resources/utils'
 import { renderToString } from '@vue/server-test-utils'
 
 describeWithShallowAndMount('options.attachToDocument', mountingMethod => {
-  it('returns VueWrapper with attachedToDocument set to true when passed attachToDocument in options', () => {
-    const compiled = compileToFunctions('<div><input /></div>')
-    const wrapper = mountingMethod(compiled, { attachToDocument: true })
+  it('attaches root node to document', () => {
+    const TestComponent = {
+      template: '<div class="attached"><input /></div>'
+    }
+    const wrapper = mountingMethod(TestComponent, {
+      attachToDocument: true
+    })
+    expect(document.querySelector('.attached')).to.not.equal(null)
     expect(wrapper.options.attachedToDocument).to.equal(true)
   })
 })
@@ -16,8 +20,10 @@ describe('options.attachToDocument with renderToString', () => {
     if (!isRunningJSDOM) {
       return
     }
-    const compiled = compileToFunctions('<div><input /></div>')
-    const fn = () => renderToString(compiled, { attachToDocument: true })
+    const TestComponent = {
+      template: '<div class="attached"><input /></div>'
+    }
+    const fn = () => renderToString(TestComponent, { attachToDocument: true })
     const message =
       '[vue-test-utils]: you cannot use attachToDocument with renderToString'
     expect(fn)