+++ test/specs/shallow-mount.spec.js
@@ -1,6 +1,6 @@
 import { compileToFunctions } from 'vue-template-compiler'
 import Vue from 'vue'
-import { mount, shallowMount } from '~vue/test-utils'
+import { mount, shallowMount, createLocalVue } from '~vue/test-utils'
 import Component from '~resources/components/component.vue'
 import ComponentWithChild from '~resources/components/component-with-child.vue'
 import ComponentWithNestedChildren from '~resources/components/component-with-nested-children.vue'
@@ -63,6 +63,44 @@ describeRunIf(process.env.TEST_ENV !== 'node', 'shallowMount', () => {
     expect(console.info.callCount).to.equal(4)
   })
 
+  it('renders children', () => {
+    const localVue = createLocalVue()
+    localVue.component('child', {
+      template: '<div />'
+    })
+    const TestComponent = {
+      template: `<child>{{'Hello'}}</child>`
+    }
+    const wrapper = shallowMount(TestComponent, {
+      localVue
+    })
+    expect(wrapper.html()).to.equal('<child-stub>Hello</child-stub>')
+  })
+
+  it('renders no children if none supplied', () => {
+    const TestComponent = {
+      template: '<child />',
+      components: { Child: { }}
+    }
+    const wrapper = shallowMount(TestComponent)
+    expect(wrapper.html()).to.equal('<child-stub></child-stub>')
+  })
+
+  it('renders children for functional components', () => {
+    const localVue = createLocalVue()
+    localVue.component('child', {
+      template: '<div />',
+      functional: true
+    })
+    const TestComponent = {
+      template: `<child>{{'Hello'}}</child>`
+    }
+    const wrapper = shallowMount(TestComponent, {
+      localVue
+    })
+    expect(wrapper.html()).to.equal('<child-stub>Hello</child-stub>')
+  })
+
   it('stubs globally registered components', () => {
     Vue.component('registered-component', ComponentWithLifecycleHooks)
     const Component = {
@@ -99,6 +137,92 @@ describeRunIf(process.env.TEST_ENV !== 'node', 'shallowMount', () => {
       expect(console.error).not.called
     }
   )
+
+  itDoNotRunIf(
+    vueVersion < 2.2, // $props does not exist in Vue < 2.2
+    'renders stubs props', () => {
+      const TestComponent = {
+        template: `<child :prop="propA" attr="hello" />`,
+        data: () => ({
+          'propA': 'a'
+        }),
+        components: {
+          child: {
+            props: ['prop']
+          }
+        }
+      }
+      const wrapper = shallowMount(TestComponent)
+      expect(wrapper.html()).to.contain('<child-stub prop="a" attr="hello"')
+    })
+
+  itDoNotRunIf(
+    vueVersion < 2.2, // $props does not exist in Vue < 2.2
+    'renders stubs classes', () => {
+      const TestComponent = {
+        template: `<child :class="classA" class="b" />`,
+        data: () => ({
+          'classA': 'a'
+        }),
+        components: {
+          child: { template: '<div />' }
+        }
+      }
+      const wrapper = shallowMount(TestComponent)
+      expect(wrapper.html()).to.contain('<child-stub class="b a"')
+    })
+
+  it('renders stubs props for functional components', () => {
+    const TestComponent = {
+      template: `<child :prop="propA" attr="hello" />`,
+      data: () => ({
+        'propA': 'a'
+      }),
+      components: {
+        Child: {
+          props: ['prop'],
+          functional: true
+        }
+      }
+    }
+    const wrapper = shallowMount(TestComponent)
+    expect(wrapper.html()).to.contain('<child-stub prop="a" attr="hello"')
+  })
+
+  it('renders classes for functional components', () => {
+    const components = {
+      Child: {
+        functional: true
+      }
+    }
+    const TestComponent = {
+      template: `<child :class="classA" class="b" />`,
+      data: () => ({
+        'classA': 'a'
+      }),
+      components
+    }
+    const wrapper = shallowMount(TestComponent)
+    expect(wrapper.html()).to.contain('<child-stub class="b a"')
+    const TestComponent2 = {
+      template: `<child :class="classA"/>`,
+      data: () => ({
+        'classA': 'a'
+      }),
+      components
+    }
+    const wrapper2 = shallowMount(TestComponent2)
+    expect(wrapper2.html()).to.contain('<child-stub class="a"')
+    const TestComponent3 = {
+      template: `<child class="b" />`,
+      data: () => ({
+        'classA': 'a'
+      }),
+      components
+    }
+    const wrapper3 = shallowMount(TestComponent3)
+    expect(wrapper3.html()).to.contain('<child-stub class="b"')
+  })
 
   itDoNotRunIf(vueVersion < 2.1, 'handles recursive components', () => {
     const TestComponent = {
@@ -254,6 +378,36 @@ describeRunIf(process.env.TEST_ENV !== 'node', 'shallowMount', () => {
 
     expect(wrapper.find(ChildComponent).vm.propA)
       .to.equal('hey')
+  })
+
+  itDoNotRunIf(
+    vueVersion < 2.4, // auto resolve of default export added in 2.4
+    'handles component as dynamic import', () => {
+      const TestComponent = {
+        template: '<div><async-component /></div>',
+        components: {
+          AsyncComponent: () => import('~resources/components/component.vue')
+        }
+      }
+      const wrapper = shallowMount(TestComponent)
+      expect(wrapper.find({ name: 'AsyncComponent' }).exists()).to.equal(true)
+    })
+
+  it('stubs components registered on localVue after multiple installs', () => {
+    const myPlugin = function (_Vue, opts) {
+      _Vue.mixin({ })
+    }
+    const localVue = createLocalVue()
+    localVue.component('registered-component', {
+      render: h => h('time')
+    })
+    const TestComponent = {
+      render: h => h('registered-component')
+    }
+
+    localVue.use(myPlugin)
+    const wrapper = shallowMount(TestComponent, { localVue })
+    expect(wrapper.html()).to.contain('registered-component-stub')
   })
 
   it('throws an error when the component fails to mount', () => {